Add role to deploy the website and Zuul job to test it
This adds a role which:
- Sets up user/group/logging for Hugo
- Retrieves and installs the hugo from upstream release
- Sets up a systemd unit file to keep the hugo server running
- Installs and configure nginx as a reverse proxy to the hugo server
  to serve ara.recordsansible.org

I'm not sure yet if keeping the hugo server running is the way to
go or if we should regenerate the static files every time.

It sort of goes against the point of hugo to use the server but
the fact that it reloads automatically when the source changes could
come in handy for automated updates.

Change-Id: I40e833a27117f9ec5003007bb08a677e16734e55