Clarify that the git commit and timestamp is for the page
Hugo retrieves the git info for each individual page.
A page that hasn't changed recently (ex: the home page) would display
an older commit and timestamp despite there being content that was
published afterwards.

Recent content have the right timestamps and git commit hashes so
this is just a clarification.

Change-Id: I785d24faca61e9411490e4271a896fd21ab2d27d