playbooks: upstream changes and move the demo to digitalocean
The website and live demo have been hosted on digitalocean for a while
but the changes had not yet been comitted because I was busy.

This brings a much needed update to the live demo playbook which still
referenced ara-web.

In a nutshell:
- It spins up a VM on digitalocean including ssh key and firewall
- It installs MySQL
- It deploys the latest version of ara with gunicorn and nginx in front

The website role needs a bit more work, it will come later.